cctview2/backend/requirements.txt
# FILE LOCATION: backend/requirements.txt

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.3.0,<3.0.0
pydantic-settings==2.1.0

# Database - Redis
redis==5.0.1
redis-om==0.3.2
hiredis==2.2.3

# Database - Neo4j
neo4j==5.14.1

# Task Queue
celery==5.3.4
kombu==5.3.4
amqp==5.2.0

# Video Processing
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
ffmpeg-python==0.2.0

# AI/ML - Vision Models
torch==2.1.1
torchvision==0.16.1
transformers==4.36.0
accelerate==0.25.0
sentencepiece==0.1.99
pillow==10.1.0

# AI/ML - Object Detection
ultralytics==8.0.228

# AI/ML - Embeddings
sentence-transformers==2.2.2
scikit-learn==1.3.2

# AI/ML - Person Re-ID
torchreid==0.2.5

# RAG/LLM
openai==1.54.3
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
httpx==0.25.2
aiofiles==23.2.1
python-dateutil==2.8.2

# Notifications
firebase-admin==6.3.0
twilio==8.11.0

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1

#Ollama
ollama